pyproject.toml
[project]
name = "fraud"
version = "0.1.0"
description = "Real-time fraud detection with XGBoost, FastAPI, and an agentic review layer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.129.0",
    "joblib>=1.5.3",
    "langchain>=1.2.10",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.8",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "psycopg[binary]>=3.2.10",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.49",
    "statsmodels>=0.14.6",
    "uvicorn>=0.41.0",
    "xgboost>=3.2.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]